refactor: move reshuffled tests back into separate files
