+++ app/src/espresso/java/xyz/lbres/trickcalculator/ui/history/testRandomness1.kt
@@ -5,15 +5,14 @@ import androidx.test.espresso.assertion.ViewAssertions.matches
 import androidx.test.espresso.matcher.ViewMatchers.isDisplayed
 import androidx.test.espresso.matcher.ViewMatchers.withText
 import org.hamcrest.Matchers.not
-import xyz.lbres.trickcalculator.R
 import xyz.lbres.trickcalculator.testutils.closeFragment
 import xyz.lbres.trickcalculator.testutils.openHistoryFragment
+import xyz.lbres.trickcalculator.testutils.textsaver.RecyclerViewTextSaver
 import xyz.lbres.trickcalculator.testutils.toggleShuffleOperators
 import xyz.lbres.trickcalculator.ui.calculator.clearText
 import xyz.lbres.trickcalculator.ui.calculator.equals
 import xyz.lbres.trickcalculator.ui.calculator.typeText
 
-private const val recyclerId = R.id.itemsRecycler
 private const val errorMessage = "History items should be shuffled in history randomness 1."
 
 fun testRandomness1() {
@@ -71,4 +70,48 @@ fun testRandomness1() {
     equals()
     computeHistory.add(TestHI("2^0.5", "Exponents must be whole numbers"))
     checkCorrectData(computeHistory, 1, errorMessage)
+}
+
+fun testRandomness1Reshuffled() {
+    setHistoryRandomness(1)
+    toggleShuffleOperators()
+
+    val computeHistory: MutableList<TestHI> = mutableListOf()
+
+    typeText("400/5")
+    equals()
+    computeHistory.add(TestHI("400/5", "80"))
+
+    clearText()
+    typeText("15-2.5")
+    equals()
+    computeHistory.add(TestHI("15-2.5", "12.5"))
+
+    clearText()
+    typeText("(3-4)(5+2)")
+    equals()
+    computeHistory.add(TestHI("(3-4)(5+2)", "-7"))
+
+    typeText("+11")
+    equals()
+    computeHistory.add(TestHI("-7+11", "4"))
+
+    clearText()
+    typeText("(3-4)(5+2)")
+    equals()
+    computeHistory.add(TestHI("(3-4)(5+2)", "-7"))
+
+    clearText()
+    typeText("+")
+    equals()
+    computeHistory.add(TestHI("+", "Syntax error"))
+
+    clearText()
+    typeText("2^0.5")
+    equals()
+    computeHistory.add(TestHI("2^0.5", "Exponents must be whole numbers"))
+
+    runSingleReshuffledTest(computeHistory, 1, errorMessage)
+    RecyclerViewTextSaver.clearAllSavedValues()
+    runSingleReshuffledTest(computeHistory, 1, errorMessage) // re-run with different order of values
 }
